% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/at.R
\name{at}
\alias{at}
\alias{at_}
\title{Format strings and escaping}
\usage{
at(.data, ...)

at_(.data, ..., .dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Format strings and escaping
}
\examples{
x <- '{"user":"stedolan","titles":["JQ Primer", "More JQ"]}'
x \%>\% at(base64) \%>\% peek
x \%>\% at(base64) \%>\% jq
x \%>\% index() \%>\% at(base64) \%>\% jq

y <- '["fo", "foo", "barfoo", "foobar", "barfoob"]'
y \%>\% index() \%>\% at(base64) \%>\% jq

## prepare for shell use
y \%>\% index() \%>\% at(sh) \%>\% jq

## rendered as csv with double quotes
z <- '[1, 2, 3, "a"]'
z \%>\% at(csv) \%>\% jq

## rendered as csv with double quotes
z \%>\% index() \%>\% jq
z \%>\% index() \%>\% at(text) \%>\% jq

## \% encode for URI's
#### DOESNT WORK --------------------------

## html escape
#### DOESNT WORK --------------------------

## serialize to json
#### DOESNT WORK --------------------------
}

