% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine json pieces}
\usage{
combine(x)
}
\arguments{
\item{x}{Input, of class json}
}
\description{
Combine json pieces
}
\examples{
x <- '{"foo": 5, "bar": 7}' \%>\% select(a = .foo)
combine(x)

(x <- githubcommits \%>\% index() \%>\%
 select(sha = .sha, name = .commit.committer.name))
combine(x)
}
