% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jqr-package.R
\docType{data}
\name{githubcommits}
\alias{githubcommits}
\title{GitHub Commits Data}
\format{A character string of json github commits data for the jq repo.}
\description{
GitHub Commits Data
}
\keyword{datasets}

