% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index_}
\alias{indexif}
\alias{indexif_}
\title{index and related functions}
\usage{
index(.data, x = "")

index_(.data, dots)

indexif(.data, x = "")

indexif_(.data, dots)
}
\arguments{
\item{.data}{input}

\item{x}{What to index to}

\item{dots}{dots}
}
\description{
index and related functions
}
\examples{
'[{"name":"JSON", "good":true}, {"name":"XML", "good":false}]' \%>\% index \%>\% jq(TRUE)
'{"a": 1, "b": 1}' \%>\% index \%>\% jq
'[]' \%>\% index \%>\% jq
'[{"name":"JSON", "good":true}, {"name":"XML", "good":false}]' \%>\% index(0) \%>\% jq
'["a","b","c","d","e"]' \%>\% index(2) \%>\% jq
'["a","b","c","d","e"]' \%>\% index(2:4) \%>\% jq
'["a","b","c","d","e"]' \%>\% index(2:5) \%>\% jq
'["a","b","c","d","e"]' \%>\% index(":3") \%>\% jq
'["a","b","c","d","e"]' \%>\% index("-2:") \%>\% jq
}

