% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jadd}
\alias{jadd}
\title{jq filter \code{add}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jadd
}
\description{
jq filter \code{add}
}
\keyword{datasets}

