% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jall}
\alias{jall}
\title{jq filter \code{all}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jall
}
\description{
jq filter \code{all}
}
\keyword{datasets}

