% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jarrays}
\alias{jarrays}
\title{jq filter \code{arrays}}
\format{\preformatted{Class 'jq'  chr "arrays"
}}
\usage{
jarrays
}
\description{
jq filter \code{arrays}
}
\keyword{datasets}

