% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jbooleans}
\alias{jbooleans}
\title{jq filter \code{booleans}}
\format{\preformatted{Class 'jq'  chr "booleans"
}}
\usage{
jbooleans
}
\description{
jq filter \code{booleans}
}
\keyword{datasets}

