% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jempty}
\alias{jempty}
\title{jq filter \code{empty}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jempty
}
\description{
jq filter \code{empty}
}
\keyword{datasets}

