% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jexplode}
\alias{jexplode}
\title{jq filter \code{explode}}
\format{\preformatted{Classes 'jq', 'character'  chr "explode"
}}
\usage{
jexplode
}
\description{
jq filter \code{explode}
}
\keyword{datasets}

