% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jfloor}
\alias{jfloor}
\title{jq filter \code{floor}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jfloor
}
\description{
jq filter \code{floor}
}
\keyword{datasets}

