% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jfrom_entries}
\alias{jfrom_entries}
\title{jq filter \code{from_entries}}
\format{\preformatted{Classes 'jq', 'character'  chr "from_entries"
}}
\usage{
jfrom_entries
}
\description{
jq filter \code{from_entries}
}
\keyword{datasets}

