% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jfromjson}
\alias{jfromjson}
\title{jq filter \code{fromjson}}
\format{\preformatted{Classes 'jq', 'character'  chr "fromjson"
}}
\usage{
jfromjson
}
\description{
jq filter \code{fromjson}
}
\keyword{datasets}

