% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jimplode}
\alias{jimplode}
\title{jq filter \code{implode}}
\format{\preformatted{Classes 'jq', 'character'  chr "implode"
}}
\usage{
jimplode
}
\description{
jq filter \code{implode}
}
\keyword{datasets}

