% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jin}
\alias{jin}
\title{jq filter \code{.}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jin
}
\description{
jq filter \code{.}
}
\keyword{datasets}

