% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jiterables}
\alias{jiterables}
\title{jq filter \code{iterables}}
\format{\preformatted{Class 'jq'  chr "iterables"
}}
\usage{
jiterables
}
\description{
jq filter \code{iterables}
}
\keyword{datasets}

