% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jkeys}
\alias{jkeys}
\title{jq filter \code{keys}}
\format{\preformatted{Classes 'jq', 'character'  chr "keys"
}}
\usage{
jkeys
}
\description{
jq filter \code{keys}
}
\keyword{datasets}

