% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jleaf_paths}
\alias{jleaf_paths}
\title{jq filter \code{leaf_paths}}
\format{An object of class \code{jq} of length 1.}
\usage{
jleaf_paths
}
\description{
jq filter \code{leaf_paths}
}
\keyword{datasets}

