% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jnulls}
\alias{jnulls}
\title{jq filter \code{nulls}}
\format{\preformatted{Classes 'jq', 'character'  chr "nulls"
}}
\usage{
jnulls
}
\description{
jq filter \code{nulls}
}
\keyword{datasets}

