% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jnumbers}
\alias{jnumbers}
\title{jq filter \code{numbers}}
\format{\preformatted{Classes 'jq', 'character'  chr "numbers"
}}
\usage{
jnumbers

jnumbers
}
\description{
jq filter \code{numbers}

jq filter \code{numbers}
}
\keyword{datasets}

