% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jobjects}
\alias{jobjects}
\title{jq filter \code{objects}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jobjects
}
\description{
jq filter \code{objects}
}
\keyword{datasets}

