% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jqr.R
\name{jq_flags}
\alias{jq_flags}
\title{Flags for use with jq}
\usage{
jq_flags(pretty = FALSE, ascii = FALSE, color = FALSE, sorted = FALSE)
}
\arguments{
\item{pretty}{Pretty print the json (different to jsonlite's
pretty printing).}

\item{ascii}{Force jq to produce pure ASCII output with non-ASCII
characters replaced by equivalent escape sequences.}

\item{color}{Add ANSI escape sequences for coloured output}

\item{sorted}{Output fields of each object with keys in sorted order}
}
\description{
Flags for use with jq
}

