% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivr.R
\name{jq_iterator}
\alias{jq_iterator}
\alias{make_jq}
\title{Create jq stream/iterator}
\usage{
jq_iterator(it, program)

make_jq(program)
}
\arguments{
\item{it}{Any iterator that yields valid jq source input}

\item{program}{JQ program, operating on whatever unit \code{it}
returns.}
}
\description{
Create stream/iterator that works by applying jq to a
stream/iterator.
}
\keyword{internal}

