% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_run.R
\name{jq_run}
\alias{jq_run}
\title{Apply a jq filter to an json input}
\usage{
jq_run(filter, json, file = NULL, args = c("--compact-output",
  "--monochrome-output"))
}
\description{
Apply a jq filter to an json input
}

