% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jreverse}
\alias{jreverse}
\title{jq filter \code{reverse}}
\format{\preformatted{Classes 'jq', 'character'  chr "reverse"
}}
\usage{
jreverse
}
\description{
jq filter \code{reverse}
}
\keyword{datasets}

