% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_run.R
\name{jrun}
\alias{jrun}
\title{Apply a jq filter to an json input}
\usage{
jrun(filter, input = jq_input(filter), args = c("--compact-output",
  "--monochrome-output"), ...)
}
\description{
Apply a jq filter to an json input
}

