% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jscalars}
\alias{jscalars}
\title{jq filter \code{scalars}}
\format{\preformatted{Classes 'jq', 'character'  chr "scalars"
}}
\usage{
jscalars
}
\description{
jq filter \code{scalars}
}
\keyword{datasets}

