% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jsqrt}
\alias{jsqrt}
\title{jq filter \code{sqrt}}
\format{\preformatted{Classes 'jq', 'character'  chr "sqrt"
}}
\usage{
jsqrt
}
\description{
jq filter \code{sqrt}
}
\keyword{datasets}

