% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jto_entries}
\alias{jto_entries}
\title{jq filter \code{to_entries}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jto_entries
}
\description{
jq filter \code{to_entries}
}
\keyword{datasets}

