% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jtojson}
\alias{jtojson}
\title{jq filter \code{tojson}}
\format{\preformatted{Classes 'jq', 'character'  chr "tojson"
}}
\usage{
jtojson
}
\description{
jq filter \code{tojson}
}
\keyword{datasets}

