% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jtostring}
\alias{jtostring}
\title{jq filter \code{tostring}}
\format{\preformatted{Classes 'jq', 'character'  chr "tostring"
}}
\usage{
jtostring
}
\description{
jq filter \code{tostring}
}
\keyword{datasets}

