% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jtype}
\alias{jtype}
\title{jq filter \code{type}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jtype
}
\description{
jq filter \code{type}
}
\keyword{datasets}

