% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jvalues}
\alias{jvalues}
\title{jq filter \code{values}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jvalues
}
\description{
jq filter \code{values}
}
\keyword{datasets}

