% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maths.R
\name{maths}
\alias{do}
\alias{do_}
\alias{floor}
\alias{length}
\alias{maths}
\alias{sqrt}
\title{Math operations}
\usage{
do(.data, ...)

do_(.data, ..., .dots)

length(.data)

sqrt(.data)

floor(.data)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Math operations
}
\examples{
\dontrun{
# do math
jqr('{"a": 7}', '.a + 1')
'{"a": 7}' \%>\%  do(.a + 1) \%>\% jq
'{"a": [1,2], "b": [3,4]}' \%>\%  do(.a + .b) \%>\% jq
'{"a": [1,2], "b": [3,4]}' \%>\%  do(.a - .b) \%>\% jq
'{"a": 3}' \%>\%  do(4 - .a) \%>\% jq
'["xml", "yaml", "json"]' \%>\%  do('. - ["xml", "yaml"]') \%>\% jq
'5' \%>\%  do(10 / . * 3) \%>\% jq

# length
'[[1,2], "string", {"a":2}, null]' \%>\% index \%>\% length \%>\% jq

# sqrt
'9' \%>\% sqrt \%>\% jq

# floor
'3.14159' \%>\% floor \%>\% jq
}
}

