% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\title{Outputs paths to all the elements in its input}
\usage{
paths(.data)
}
\arguments{
\item{.data}{input}
}
\description{
Outputs paths to all the elements in its input
}
\examples{
'[1,[[],{"a":2}]]' \%>\% paths
'[{"name":"JSON", "good":true}, {"name":"XML", "good":false}]' \%>\% paths
}

