% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Peek at a query}
\usage{
peek(.data)
}
\arguments{
\item{.data}{(list) input, using higher level interface}
}
\description{
Prints the query resulting from \code{jq} all in one character string just
as you would execute it on the command line. Output gets class of json,
and pretty prints to the console for easier viewing. The \code{jqr}
doesn't do pretty printing.
}
\examples{
'{"a": 7}' \%>\%  do(.a + 1) \%>\% peek
'[8,3,null,6]' \%>\% sortj \%>\% peek
}
\seealso{
\code{\link{jq}}, \code{\link{jq_}}
}

