% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/range.R
\name{rangej}
\alias{rangej}
\title{Produce range of numbers}
\usage{
rangej(x, array = FALSE)
}
\arguments{
\item{x}{Input, single number or number range.}

\item{array}{(logical) Create array. Deafult: FALSE}
}
\description{
Produce range of numbers
}
\examples{
rangej(2:4)
rangej(2:4) \%>\% jq
rangej(2:4, TRUE)
rangej(2:4, TRUE) \%>\% jq
# not working for some reason
# rangej(4, TRUE) \%>\% jq
}

