% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurse.R
\name{recurse}
\alias{recurse}
\alias{recurse_}
\title{Search through a recursive structure, and extract interesting data from all levels}
\usage{
recurse(.data, ...)

recurse_(.data, ..., .dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Search through a recursive structure, and extract interesting data from all levels
}

