% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{string}
\alias{string}
\title{Give back a character string}
\usage{
string(.data)
}
\arguments{
\item{.data}{(list) input, using higher level interface}
}
\description{
Give back a character string
}
\examples{
'{"a": 7}' \%>\% do(.a + 1) \%>\% string
'[8,3,null,6]' \%>\% sortj \%>\% string
}
\seealso{
\code{\link{peek}}
}

