% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at.R
\name{at}
\alias{at}
\alias{at_}
\title{Format strings and escaping}
\usage{
at(.data, ...)

at_(.data, ..., .dots)
}
\arguments{
\item{.data}{input. This can be JSON input, or an object of class
\code{jqr} that has JSON and query params combined, which is passed
from function to function when using the jqr DSL.}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Format strings and escaping
}
\examples{
x <- '{"user":"stedolan","titles":["JQ Primer", "More JQ"]}'
# x \%>\% at(base64) \%>\% peek
x \%>\% at(base64)
x \%>\% index() \%>\% at(base64)

y <- '["fo", "foo", "barfoo", "foobar", "barfoob"]'
y \%>\% index() \%>\% at(base64)

## prepare for shell use
y \%>\% index() \%>\% at(sh)

## rendered as csv with double quotes
z <- '[1, 2, 3, "a"]'
z \%>\% at(csv)

## rendered as csv with double quotes
z \%>\% index()
z \%>\% index() \%>\% at(text)

## \% encode for URI's
#### DOESNT WORK --------------------------

## html escape
#### DOESNT WORK --------------------------

## serialize to json
#### DOESNT WORK --------------------------
}
