% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dot.R
\name{dot}
\alias{dot}
\alias{dot_}
\alias{dotstr}
\alias{dotstr_}
\title{dot and related functions}
\usage{
dot(.data)

dot_(.data, dots)

dotstr(.data, ...)

dotstr_(.data, ..., .dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
dot and related functions
}

