% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index_}
\alias{indexif}
\alias{indexif_}
\title{index and related functions}
\usage{
index(.data, x = "")

index_(.data, dots)

indexif(.data, x = "")

indexif_(.data, dots)
}
\arguments{
\item{.data}{input}

\item{x}{What to index to}

\item{dots}{dots}
}
\description{
index and related functions
}
\examples{
str <- '[{"name":"JSON", "good":true}, {"name":"XML", "good":false}]'
str \%>\% index
str \%>\% indexif
str \%>\% indexif('that')
'{"a": 1, "b": 1}' \%>\% index
'[]' \%>\% index
'[{"name":"JSON", "good":true}, {"name":"XML", "good":false}]' \%>\% index(0)
'["a","b","c","d","e"]' \%>\% index(2)
'["a","b","c","d","e"]' \%>\% index(2:4)
'["a","b","c","d","e"]' \%>\% index(2:5)
'["a","b","c","d","e"]' \%>\% index(":3")
'["a","b","c","d","e"]' \%>\% index("-2:")

str \%>\% index \%>\% select(bad = .name)
}

