% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index_}
\alias{indexif}
\alias{indexif_}
\title{index and related functions}
\usage{
index(.data, x = "")

index_(.data, dots)

indexif(.data, x = "")

indexif_(.data, dots)
}
\arguments{
\item{.data}{input}

\item{x}{What to index to}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
index and related functions
}

