% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jarrays}
\alias{jarrays}
\title{jq filter \code{arrays}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jarrays
}
\description{
jq filter \code{arrays}
}
\keyword{datasets}

