% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jbooleans}
\alias{jbooleans}
\title{jq filter \code{booleans}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jbooleans
}
\description{
jq filter \code{booleans}
}
\keyword{datasets}

