% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jempty}
\alias{jempty}
\title{jq filter \code{empty}}
\format{\preformatted{Classes 'jq', 'character'  chr "empty"
}}
\usage{
jempty
}
\description{
jq filter \code{empty}
}
\keyword{datasets}

