% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jexplode}
\alias{jexplode}
\title{jq filter \code{explode}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jexplode
}
\description{
jq filter \code{explode}
}
\keyword{datasets}

