% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jfrom_entries}
\alias{jfrom_entries}
\title{jq filter \code{from_entries}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jfrom_entries
}
\description{
jq filter \code{from_entries}
}
\keyword{datasets}

