% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jfromjson}
\alias{jfromjson}
\title{jq filter \code{fromjson}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jfromjson
}
\description{
jq filter \code{fromjson}
}
\keyword{datasets}

