% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jiterables}
\alias{jiterables}
\title{jq filter \code{iterables}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jiterables
}
\description{
jq filter \code{iterables}
}
\keyword{datasets}

