% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jkeys}
\alias{jkeys}
\title{jq filter \code{keys}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jkeys
}
\description{
jq filter \code{keys}
}
\keyword{datasets}

