% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jleaf_paths}
\alias{jleaf_paths}
\title{jq filter \code{leaf_paths}}
\format{\preformatted{Classes 'jq', 'character'  chr "leaf_paths"
}}
\usage{
jleaf_paths
}
\description{
jq filter \code{leaf_paths}
}
\keyword{datasets}

