% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jlength}
\alias{jlength}
\title{jq filter \code{length}}
\format{\preformatted{Classes 'jq', 'character'  chr "length"
}}
\usage{
jlength
}
\description{
jq filter \code{length}
}
\keyword{datasets}

