% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jlength}
\alias{jlength}
\title{jq filter \code{length}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jlength
}
\description{
jq filter \code{length}
}
\keyword{datasets}

