% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jnulls}
\alias{jnulls}
\title{jq filter \code{nulls}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jnulls
}
\description{
jq filter \code{nulls}
}
\keyword{datasets}

