% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jnumbers}
\alias{jnumbers}
\title{jq filter \code{numbers}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jnumbers

jnumbers
}
\description{
jq filter \code{numbers}

jq filter \code{numbers}
}
\keyword{datasets}

