% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jobjects}
\alias{jobjects}
\title{jq filter \code{objects}}
\format{\preformatted{Classes 'jq', 'character'  chr "objects"
}}
\usage{
jobjects
}
\description{
jq filter \code{objects}
}
\keyword{datasets}

