% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jqr.R
\name{jq}
\alias{jq}
\alias{jq_}
\title{Execute a query with jq}
\usage{
jq(.data, pretty = FALSE)

jq_(data, query)
}
\arguments{
\item{.data}{(list) input, using higher level interface}

\item{pretty}{(logical) Add newlines when printed. Default: FALSE}

\item{data}{(character) JSON data}

\item{query}{(character) A query string}
}
\description{
\code{jq} is meant to work with the high level interface in this package. \code{jq_}
provides access to the low level interface in which you can use jq query strings just
as you would on the command line. Output gets class of json, and pretty prints to
the console for easier viewing. The \code{jqr} doesn't do pretty printing.
}
\examples{
'{"a": 7}' \%>\%  do(.a + 1) \%>\% jq
'[8,3,null,6]' \%>\% sortj \%>\% jq
}
\seealso{
\code{\link{peek}}
}

