% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jq}
\alias{jq}
\title{parse json}
\usage{
jq(json, ...)
}
\arguments{
\item{json}{(character) Input json string}

\item{...}{Addtn inputs to fromJSON}
}
\description{
parse json
}
\examples{
library("httr")
library("magrittr")
library("data.table")

# GBIF example
json <- content(GET('http://api.gbif.org/v0.9/occurrence/search?taxonKey=3119195'), as = "text")
json \%>\%
   jq \%s\%
   'results' \%v\%
   c('scientificName','country') \%>\%
   rbindlist

# PLOS example
json <- content(GET('http://api.plos.org/search/?q=*:*&wt=json'), as = "text")
json \%>\%
   jq \%s\%
   'response.docs' \%v\%
   c('id','journal') \%>\%
   rbindlist
}

