% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jqr-package.R
\docType{package}
\name{jqr}
\alias{jqr}
\alias{jqr-package}
\title{jqr: An R client for the C library jq}
\description{
jqr: An R client for the C library jq
}
\section{Low-level}{

Low level interface, in which you can execute `jq` code just as you
would on the command line. Available via \code{\link{jq_}}
}

\section{High-level DSL}{

High-level, uses a suite of functions to construct queries. Queries
are constucted, then excuted internally with `jq()`.
}

\section{Pipes}{

The high level DSL supports piping, though you don't have to use
pipes.
}

\section{jq version}{

\code{jqr} is currenlty using \code{jq v1.4}
}

