% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jqr-package.R
\docType{package}
\name{jqr}
\alias{jqr}
\alias{jqr-package}
\title{jqr: An R client for the C library jq}
\description{
\code{\link{jq}} and \code{\link{jq_}} are the most important functions.
\code{\link{jq}} is meant to work with the high level interface in this
package. \code{\link{jq_}} provides access to the low level interface
in which you can use jq query strings just as you would on the command
line.
}

