% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jscalars}
\alias{jscalars}
\title{jq filter \code{scalars}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jscalars
}
\description{
jq filter \code{scalars}
}
\keyword{datasets}

