% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jsqrt}
\alias{jsqrt}
\title{jq filter \code{sqrt}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jsqrt
}
\description{
jq filter \code{sqrt}
}
\keyword{datasets}

