% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jstrings}
\alias{jstrings}
\title{jq filter \code{strings}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jstrings
}
\description{
jq filter \code{strings}
}
\keyword{datasets}

