% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jstrings}
\alias{jstrings}
\title{jq filter \code{strings}}
\format{\preformatted{Classes 'jq', 'character'  chr "strings"
}}
\usage{
jstrings
}
\description{
jq filter \code{strings}
}
\keyword{datasets}

