% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jto_entries}
\alias{jto_entries}
\title{jq filter \code{to_entries}}
\format{\preformatted{Class 'jq'  chr "to_entries"
}}
\usage{
jto_entries
}
\description{
jq filter \code{to_entries}
}
\keyword{datasets}

