% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jtonumber}
\alias{jtonumber}
\title{jq filter \code{tonumber}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jtonumber
}
\description{
jq filter \code{tonumber}
}
\keyword{datasets}

