% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jtonumber}
\alias{jtonumber}
\title{jq filter \code{tonumber}}
\format{\preformatted{Classes 'jq', 'character'  chr "tonumber"
}}
\usage{
jtonumber
}
\description{
jq filter \code{tonumber}
}
\keyword{datasets}

