% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jtostring}
\alias{jtostring}
\title{jq filter \code{tostring}}
\format{An object of class \code{jq} (inherits from \code{character}) of length 1.}
\usage{
jtostring
}
\description{
jq filter \code{tostring}
}
\keyword{datasets}

