% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jtype}
\alias{jtype}
\title{jq filter \code{type}}
\format{\preformatted{Classes 'jq', 'character'  chr "type"
}}
\usage{
jtype
}
\description{
jq filter \code{type}
}
\keyword{datasets}

