% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jq_dsl.R
\docType{data}
\name{jvalues}
\alias{jvalues}
\title{jq filter \code{values}}
\format{\preformatted{Class 'jq'  chr "values"
}}
\usage{
jvalues
}
\description{
jq filter \code{values}
}
\keyword{datasets}

