% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logicaltests.R
\name{logicaltests}
\alias{all}
\alias{any}
\alias{logicaltests}
\title{Logical tests}
\usage{
all(.data)

any(.data)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Logical tests
}
\examples{
# any
'[true, false]' \%>\% any \%>\% jq
'[false, false]' \%>\% any \%>\% jq
'[]' \%>\% any \%>\% jq

# all
'[true, false]' \%>\% all \%>\% jq
'[true, true]' \%>\% all \%>\% jq
'[]' \%>\% add \%>\% jq
}

