% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logicaltests.R
\name{logicaltests}
\alias{allj}
\alias{anyj}
\alias{logicaltests}
\title{Logical tests}
\usage{
allj(.data)

anyj(.data)
}
\arguments{
\item{.data}{input}
}
\description{
Logical tests
}
\examples{
# any
'[true, false]' \%>\% anyj \%>\% jq
'[false, false]' \%>\% anyj \%>\% jq
'[]' \%>\% anyj \%>\% jq

# all
'[true, false]' \%>\% allj \%>\% jq
'[true, true]' \%>\% allj \%>\% jq
'[]' \%>\% allj \%>\% jq
}

