% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicaltests.R
\name{logicaltests}
\alias{allj}
\alias{anyj}
\alias{logicaltests}
\title{Logical tests}
\usage{
allj(.data)

anyj(.data)
}
\arguments{
\item{.data}{input. This can be JSON input, or an object of class
\code{jqr} that has JSON and query params combined, which is passed
from function to function when using the jqr DSL.}
}
\description{
Logical tests
}
\examples{
# any
'[true, false]' \%>\% anyj
'[false, false]' \%>\% anyj
'[]' \%>\% anyj

# all
'[true, false]' \%>\% allj
'[true, true]' \%>\% allj
'[]' \%>\% allj
}

