% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manip.R
\name{manip}
\alias{contains}
\alias{contains_}
\alias{endswith}
\alias{endswith_}
\alias{fromjson}
\alias{group}
\alias{group_}
\alias{index_loc}
\alias{index_loc_}
\alias{indices}
\alias{indices_}
\alias{join}
\alias{join_}
\alias{ltrimstr}
\alias{ltrimstr_}
\alias{manip}
\alias{rindex_loc}
\alias{rindex_loc_}
\alias{rtrimstr}
\alias{rtrimstr_}
\alias{splitj}
\alias{splitj_}
\alias{startswith}
\alias{startswith_}
\alias{tojson}
\alias{tostring}
\alias{uniquej}
\alias{uniquej_}
\title{Manipulation operations}
\usage{
join(.data, ...)

join_(.data, ..., .dots)

splitj(.data, ...)

splitj_(.data, ..., .dots)

ltrimstr(.data, ...)

ltrimstr_(.data, ..., .dots)

rtrimstr(.data, ...)

rtrimstr_(.data, ..., .dots)

startswith(.data, ...)

startswith_(.data, ..., .dots)

endswith(.data, ...)

endswith_(.data, ..., .dots)

index_loc(.data, ...)

index_loc_(.data, ..., .dots)

rindex_loc(.data, ...)

rindex_loc_(.data, ..., .dots)

indices(.data, ...)

indices_(.data, ..., .dots)

tojson(.data)

fromjson(.data)

tostring(.data)

contains(.data, ...)

contains_(.data, ..., .dots)

uniquej(.data, ...)

uniquej_(.data, ..., .dots)

group(.data, ...)

group_(.data, ..., .dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Manipulation operations
}
\examples{
# join
jq_('["a","b,c,d","e"]', 'join(", ")')
'["a","b,c,d","e"]' \%>\% join \%>\% jq
'["a","b,c,d","e"]' \%>\% join(`;`) \%>\% jq
'["a","b,c,d","e"]' \%>\% join(`yep`) \%>\% jq

# split
jq_('"a, b,c,d, e"', 'split(", ")')

# ltrimstr
jq_('["fo", "foo", "barfoo", "foobar", "afoo"]', '[.[]|ltrimstr("foo")]')
'["fo", "foo", "barfoo", "foobar", "afoo"]' \%>\% index() \%>\% ltrimstr(foo) \%>\% jq

# rtrimstr
jq_('["fo", "foo", "barfoo", "foobar", "foob"]', '[.[]|rtrimstr("foo")]')
'["fo", "foo", "barfoo", "foobar", "foob"]' \%>\% index() \%>\% rtrimstr(foo) \%>\% jq

# startswith
jq_('["fo", "foo", "barfoo", "foobar", "barfoob"]', '[.[]|startswith("foo")]')
'["fo", "foo", "barfoo", "foobar", "barfoob"]' \%>\% index \%>\% startswith(foo) \%>\% jq

# endswith
jq_('["fo", "foo", "barfoo", "foobar", "barfoob"]', '[.[]|endswith("foo")]')
'["fo", "foo", "barfoo", "foobar", "barfoob"]' \%>\% index \%>\% endswith(foo) \%>\% jq
'["fo", "foo", "barfoo", "foobar", "barfoob"]' \%>\% index \%>\% endswith(bar) \%>\% jq

# get index (location) of a character
## input has to be quoted
'"a,b, cd, efg, hijk"' \%>\% index_loc(", ") \%>\% jq
'"a,b, cd, efg, hijk"' \%>\% index_loc(",") \%>\% jq
'"a,b, cd, efg, hijk"' \%>\% index_loc("j") \%>\% jq
'"a,b, cd, efg, hijk"' \%>\% rindex_loc(", ") \%>\% jq
'"a,b, cd, efg, hijk"' \%>\% indices(", ") \%>\% jq

# tojson, fromjson, tostring
'[1, "foo", ["foo"]]' \%>\% index \%>\% tostring \%>\% jq
'[1, "foo", ["foo"]]' \%>\% index \%>\% tojson \%>\% jq
'[1, "foo", ["foo"]]' \%>\% index \%>\% tojson \%>\% fromjson \%>\% jq

# contains
'"foobar"' \%>\% contains("bar") \%>\% jq
'["foobar", "foobaz", "blarp"]' \%>\% contains(`["baz", "bar"]`) \%>\% jq
'["foobar", "foobaz", "blarp"]' \%>\% contains(`["bazzzzz", "bar"]`) \%>\% jq
str <- '{"foo": 12, "bar":[1,2,{"barp":12, "blip":13}]}'
str \%>\% contains(`{foo: 12, bar: [{barp: 12}]}`) \%>\% jq
str \%>\% contains(`{foo: 12, bar: [{barp: 15}]}`) \%>\% jq

# unique
'[1,2,5,3,5,3,1,3]' \%>\% uniquej \%>\% jq
str <- '[{"foo": 1, "bar": 2}, {"foo": 1, "bar": 3}, {"foo": 4, "bar": 5}]'
str \%>\% uniquej(foo) \%>\% jq
'["chunky", "bacon", "kitten", "cicada", "asparagus"]' \%>\% uniquej(length) \%>\% jq

# group
x <- '[{"foo":1, "bar":10}, {"foo":3, "bar":100}, {"foo":1, "bar":1}]'
x \%>\% group(foo)
x \%>\% group(foo) \%>\% jq
}
\seealso{
\code{\link{add}}
}

