% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maths.R
\name{maths}
\alias{length}
\alias{length_}
\alias{maths}
\title{Math operations}
\usage{
length(.data)

length_(.data, dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}
}
\description{
Math operations
}

