% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maths.R
\name{maths}
\alias{add}
\alias{do}
\alias{do_}
\alias{floorj}
\alias{lengthj}
\alias{map}
\alias{map_}
\alias{maths}
\alias{maxj}
\alias{maxj_}
\alias{minj}
\alias{minj_}
\alias{sqrtj}
\title{Math operations}
\usage{
do(.data, ...)

do_(.data, ..., .dots)

lengthj(.data)

sqrtj(.data)

floorj(.data)

minj(.data, ...)

minj_(.data, ..., .dots)

maxj(.data, ...)

maxj_(.data, ..., .dots)

add(.data)

map(.data, ...)

map_(.data, ..., .dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Math operations
}
\examples{
# do math
jq_('{"a": 7}', '.a + 1')
# adding null gives back same result
jq_('{"a": 7}', '.a + null')
jq_('{"a": 7}', '.a += 1')
'{"a": 7}' \%>\%  do(.a + 1) \%>\% jq
# '{"a": 7}' \%>\%  do(.a += 1) \%>\% jq # this doesn't work quite yet
'{"a": [1,2], "b": [3,4]}' \%>\%  do(.a + .b) \%>\% jq
'{"a": [1,2], "b": [3,4]}' \%>\%  do(.a - .b) \%>\% jq
'{"a": 3}' \%>\%  do(4 - .a) \%>\% jq
'["xml", "yaml", "json"]' \%>\%  do('. - ["xml", "yaml"]') \%>\% jq
'5' \%>\%  do(10 / . * 3) \%>\% jq

# comparisons
'[5,4,2,7]' \%>\% index() \%>\% do(. < 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. > 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. <= 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. >= 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. == 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. != 4) \%>\% jq

# length
'[[1,2], "string", {"a":2}, null]' \%>\% index \%>\% lengthj \%>\% jq

# sqrt
'9' \%>\% sqrtj \%>\% jq

# floor
'3.14159' \%>\% floorj \%>\% jq

# find minimum
'[5,4,2,7]' \%>\% minj \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% minj \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% minj(foo) \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% minj(bar) \%>\% jq

# find maximum
'[5,4,2,7]' \%>\% maxj \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% maxj \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% maxj(foo) \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% maxj(bar) \%>\% jq

# increment values
## requires special \% operators, they get escaped internally
'{"foo": 1}' \%>\% do(.foo \%+=\% 1) \%>\% jq
'{"foo": 1}' \%>\% do(.foo \%-=\% 1) \%>\% jq
'{"foo": 1}' \%>\% do(.foo \%*=\% 4) \%>\% jq
'{"foo": 1}' \%>\% do(.foo \%/=\% 10) \%>\% jq
'{"foo": 1}' \%>\% do(.foo \%//=\% 10) \%>\% jq
### fix me - \%= doesn't work
# '{"foo": 1}' \%>\% do(.foo \%\%=\% 10) \%>\% jq

# add
'["a","b","c"]' \%>\% add \%>\% jq
'[1, 2, 3]' \%>\% add \%>\% jq
'[]' \%>\% add \%>\% jq

# map
## as far as I know, this only works with numbers, thus it's
## in the maths section
'[1, 2, 3]' \%>\% map(.+1) \%>\% jq
'[1, 2, 3]' \%>\% map(./1) \%>\% jq
'[1, 2, 3]' \%>\% map(.*4) \%>\% jq
}

