% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maths.R
\name{maths}
\alias{do}
\alias{do_}
\alias{floor}
\alias{length}
\alias{maths}
\alias{max}
\alias{max_}
\alias{min}
\alias{min_}
\alias{sqrt}
\title{Math operations}
\usage{
do(.data, ...)

do_(.data, ..., .dots)

length(.data)

sqrt(.data)

floor(.data)

min(.data, ...)

min_(.data, ..., .dots)

max(.data, ...)

max_(.data, ..., .dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Math operations
}
\examples{
\dontrun{
# do math
jqr('{"a": 7}', '.a + 1')
jqr('{"a": 7}', '.a += 1')
'{"a": 7}' \%>\%  do(.a + 1) \%>\% jq
# '{"a": 7}' \%>\%  do(.a += 1) \%>\% jq # this doesn't work quite yet
'{"a": [1,2], "b": [3,4]}' \%>\%  do(.a + .b) \%>\% jq
'{"a": [1,2], "b": [3,4]}' \%>\%  do(.a - .b) \%>\% jq
'{"a": 3}' \%>\%  do(4 - .a) \%>\% jq
'["xml", "yaml", "json"]' \%>\%  do('. - ["xml", "yaml"]') \%>\% jq
'5' \%>\%  do(10 / . * 3) \%>\% jq
## comparisons
'[5,4,2,7]' \%>\% index() \%>\% do(. < 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. > 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. <= 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. >= 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. == 4) \%>\% jq
'[5,4,2,7]' \%>\% index() \%>\% do(. != 4) \%>\% jq

# length
'[[1,2], "string", {"a":2}, null]' \%>\% index \%>\% length \%>\% jq

# sqrt
'9' \%>\% sqrt \%>\% jq

# floor
'3.14159' \%>\% floor \%>\% jq

# find minimum
'[5,4,2,7]' \%>\% min \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% min \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% min(foo) \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% min(bar) \%>\% jq

# find maximum
'[5,4,2,7]' \%>\% max \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% max \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% max(foo) \%>\% jq
'[{"foo":1, "bar":14}, {"foo":2, "bar":3}]' \%>\% max(bar) \%>\% jq
}
}

