% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{rangej}
\alias{rangej}
\title{Produce range of numbers}
\usage{
rangej(x, array = FALSE)
}
\arguments{
\item{x}{Input, single number or number range.}

\item{array}{(logical) Create array. Default: \code{FALSE}}
}
\description{
Produce range of numbers
}
\examples{
rangej(2:4)
2:4 \%>\% rangej
2:1000 \%>\% rangej
# errors, correctly though
# 2 \%>\% rangej
}
