% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\alias{select_}
\title{Select variables}
\usage{
select(.data, ...)

select_(.data, ..., .dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Select variables
}
\examples{
'{"foo": 5, "bar": 7}' \%>\% select(a = .foo) \%>\% peek
'{"foo": 5, "bar": 7}' \%>\% select(a = .foo) \%>\% jq

# using json dataset, just first element
x <- githubcommits \%>\% index(0)
x \%>\%
   select(message = .commit.message, name = .commit.committer.name) \%>\% jq
x \%>\%
   select(sha = .commit.tree.sha, author = .author.login) \%>\% jq

# using json dataset, all elements
x <- githubcommits \%>\% index()
x \%>\%
   select(message = .commit.message, name = .commit.committer.name) \%>\% jq
## pretty (newline after each element)
x \%>\%
 select(sha = .sha, name = .commit.committer.name) \%>\% jq(TRUE)
}

