% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\alias{select_}
\title{select variables}
\usage{
select(.data, ...)

select_(.data, ..., .dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
select variables
}

