% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/types.R
\name{types}
\alias{type}
\alias{type_}
\alias{types}
\title{Types and related functions}
\usage{
types(.data)

type(.data, ...)

type_(.data, ..., .dots)
}
\arguments{
\item{.data}{input}

\item{...}{Comma separated list of unquoted variable names}

\item{.dots}{Used to work around non-standard evaluation}

\item{dots}{dots}
}
\description{
Types and related functions
}
\examples{
\dontrun{
# get type information for each element
jqr('[0, false, [], {}, null, "hello"]', 'map(type)')
'[0, false, [], {}, null, "hello"]' \%>\% types \%>\% jq
'[0, false, [], {}, null, "hello", true, [1,2,3]]' \%>\% types \%>\% jq

# select elements by type
jqr('[0, false, [], {}, null, "hello"]', '.[] | numbers,booleans')
'[0, false, [], {}, null, "hello"]' \%>\% index() \%>\% type(booleans) \%>\% jq
}
}

